디렉터리(패키지) 구조 분류, 대분류화
