+++ src/main/java/com/antmillion/communication/dto/CommunityDTO.java
@@ -1,4 +1,4 @@
-package com.antmillion.dto;
+package com.antmillion.communication.dto;
 
 import java.time.LocalDateTime;
 
@@ -12,12 +12,10 @@ import lombok.ToString;
 @NoArgsConstructor
 @AllArgsConstructor
 @ToString
-public class HistoryDTO {
-	private Long history_id;
+public class CommunityDTO {
+	private Long post_id;
     private Long user_id;
-    private Long message_id;
     private String stock_code;
-    private String message_detail;
-    private LocalDateTime time;
-    private Boolean is_read;
+    private String content;
+    private LocalDateTime posted_date;
 }
