+++ src/main/java/com/antmillion/communication/dto/MessageDTO.java
@@ -1,4 +1,4 @@
-package com.antmillion.dto;
+package com.antmillion.communication.dto;
 
 import lombok.AllArgsConstructor;
 import lombok.Getter;
@@ -10,9 +10,8 @@ import lombok.ToString;
 @NoArgsConstructor
 @AllArgsConstructor
 @ToString
-public class AccountDTO {
-	 private Long account_id;
-	    private Long user_id;
-	    private String account_number;
-	    private Long balance;
+public class MessageDTO {
+	private Long message_id;
+    private String message_content;
+    private String message_type;
 }
