--- src/main/java/com/antmillion/dto/MissionDTO.java
@@ -1,18 +1,0 @@
-package com.antmillion.dto;
-
-import java.math.BigDecimal;
-
-import lombok.AllArgsConstructor;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-import lombok.Setter;
-import lombok.ToString;
-
-@Getter@Setter
-@NoArgsConstructor
-@AllArgsConstructor
-@ToString
-public class MissionDTO {
-	private Long mission_id;
-    private String mission_name;
-}
