--- src/main/java/com/antmillion/dto/StockDTO.java
@@ -1,17 +1,0 @@
-package com.antmillion.dto;
-
-import lombok.AllArgsConstructor;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-import lombok.Setter;
-import lombok.ToString;
-
-@Getter@Setter
-@NoArgsConstructor
-@AllArgsConstructor
-@ToString
-public class StockDTO {
-	private String stock_code;
-	private String stock_name;
-	private String stock_image;
-}
