+++ src/main/java/com/antmillion/mypage/controller/MyPageController.java
@@ -1,16 +1,16 @@
-package com.antmillion.main;
+package com.antmillion.mypage.controller;
 
 import org.springframework.stereotype.Controller;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.RequestMapping;
 
-@RequestMapping({"/", "/antmillion", "/antmillion/"})
+@RequestMapping({"/mypage"})
 @Controller
-public class MainController {
+public class MyPageController {
 
     @GetMapping
     public String mainPage() {
-        return "main/main";
+        return "mypage/mypage";
     }
 
 }
