+++ src/main/java/com/antmillion/stock/controller/StockDetailController.java
@@ -1,4 +1,4 @@
-package com.antmillion.stock;
+package com.antmillion.stock.controller;
 
 import org.springframework.stereotype.Controller;
 import org.springframework.web.bind.annotation.GetMapping;
