+++ src/main/java/com/antmillion/stock/controller/StockListController.java
@@ -1,15 +1,16 @@
-package com.antmillion.mission;
+package com.antmillion.stock.controller;
 
 import org.springframework.stereotype.Controller;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.RequestMapping;
 
+@RequestMapping("/stocklist")
 @Controller
-@RequestMapping("/mission")
-public class MissionController {
+public class StockListController {
 
     @GetMapping
-    public String missionPage() {
-        return "mission/mission";
+    public String stockListPage() {
+        return "stocklist/stocklist";
     }
+
 }
