+++ src/main/java/com/antmillion/temp/controller/HIstoryController.java
@@ -1,4 +1,4 @@
-package com.antmillion.history;
+package com.antmillion.temp.controller;
 
 import org.springframework.stereotype.Controller;
 import org.springframework.web.bind.annotation.GetMapping;
