+++ src/main/java/com/antmillion/user/controller/LoginController.java
@@ -1,4 +1,4 @@
-package com.antmillion.login;
+package com.antmillion.user.controller;
 
 import org.springframework.stereotype.Controller;
 import org.springframework.web.bind.annotation.GetMapping;
