+++ src/main/java/com/antmillion/user/dto/AntRankDTO.java
@@ -1,6 +1,4 @@
-package com.antmillion.dto;
-
-import java.time.LocalDateTime;
+package com.antmillion.user.dto;
 
 import lombok.AllArgsConstructor;
 import lombok.Getter;
@@ -12,8 +10,9 @@ import lombok.ToString;
 @NoArgsConstructor
 @AllArgsConstructor
 @ToString
-public class InterestDTO {
-	private String stock_code;
-    private Long account_id;
-    private LocalDateTime created_at;
+public class AntRankDTO {
+    private Integer rank_id;
+    private String rank_type;
+    private Integer required_point;
+    private String rank_image;
 }
